Create Email exempted by policy
This query can help identify mail that may have been exempted by settings in the IP/Domain/Sender safelists. This can be used in conjunction with the ORCA report to hunt and query for email that is specifically bypassing filtering due to the fact the sender is on the EOP safe list as defined by the customer/tenant owner.